pyproject.toml

[project]
name = "imagecaption"
version = "0.0.1"
description = "Image captioning using VIT and GPT2"
authors = [
  { name = "Harnoor Rangi", email = "harnoorrangi@hotmail.com" },
]

readme = "README.md"
requires-python = "==3.10.14"
dependencies = [
    "cmake>=3.31.1",
    "datasets>=3.2.0",
    "evaluate>=0.4.3",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "matplotlib>=3.9.3",
    "numpy<2",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pip>=24.3.1",
    "pyarrow<18.0.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
    "torch==2.0.1",
    "torchvision>=0.15.2",
    "transformers>=4.46.3",
    "typing>=3.10.0.0",
]

[project.scripts]
example-pkg = "example_packaged_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
pre-commit = ["pre-commit >= 3.5.0"]
testing = ["pytest >= 7.4.3"]
linting = ["ruff >= 0.1.3"]
docs = [
  "mkdocs <= 1.5.3",
  "mkdocs-material <= 9.4.6",
  "mkdocstrings <= 0.23.0",
  "mkdocstrings[python] >= 0.18"
]

[tool.ruff]
line-length = 120
